feat: Adiciona a bluprint pratos e docstrings
